scripts/petri_net_generation/requirements.txt
pm4py
PyYAML